phospho-mcp-server/pyproject.toml
[project]
name = "phospho-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]==1.10.0",
    "pillow",
    "requests",
    "opencv-python",
]

[tool.uv.workspace]
members = ["phospho-mcp-server"]
